requirements.txt
googlesearch-python==1.0.2